Add gx.Track to MultiGeometry and adjust import
