+++ fastkml/features.py
@@ -30,7 +30,6 @@ from pygeoif.types import GeoType
 
 from fastkml import atom
 from fastkml import config
-from fastkml import gx
 from fastkml.base import _XMLObject
 from fastkml.data import ExtendedData
 from fastkml.geometry import AnyGeometryType
@@ -40,6 +39,8 @@ from fastkml.geometry import MultiGeometry
 from fastkml.geometry import Point
 from fastkml.geometry import Polygon
 from fastkml.geometry import create_kml_geometry
+from fastkml.gx.track import MultiTrack
+from fastkml.gx.track import Track
 from fastkml.helpers import attribute_int_kwarg
 from fastkml.helpers import bool_subelement
 from fastkml.helpers import clean_string
@@ -79,8 +80,8 @@ KmlGeometry = Union[
     Polygon,
     Model,
     MultiGeometry,
-    gx.MultiTrack,
-    gx.Track,
+    MultiTrack,
+    Track,
 ]
 
 
@@ -675,8 +676,8 @@ registry.register(
             Polygon,
             MultiGeometry,
             Model,
-            gx.MultiTrack,
-            gx.Track,
+            MultiTrack,
+            Track,
         ),
         get_kwarg=xml_subelement_kwarg,
         set_element=xml_subelement,
