+++ fastkml/overlays.py
@@ -23,7 +23,6 @@ from typing import Union
 
 from fastkml import atom
 from fastkml import config
-from fastkml import gx
 from fastkml.base import _XMLObject
 from fastkml.data import ExtendedData
 from fastkml.enums import AltitudeMode
@@ -32,11 +31,7 @@ from fastkml.enums import Shape
 from fastkml.enums import Units
 from fastkml.features import Snippet
 from fastkml.features import _Feature
-from fastkml.geometry import LinearRing
-from fastkml.geometry import LineString
-from fastkml.geometry import MultiGeometry
 from fastkml.geometry import Point
-from fastkml.geometry import Polygon
 from fastkml.helpers import attribute_enum_kwarg
 from fastkml.helpers import attribute_float_kwarg
 from fastkml.helpers import clean_string
@@ -79,16 +74,6 @@ __all__ = [
 ]
 
 logger = logging.getLogger(__name__)
-
-KmlGeometry = Union[
-    Point,
-    LineString,
-    LinearRing,
-    Polygon,
-    MultiGeometry,
-    gx.MultiTrack,
-    gx.Track,
-]
 
 
 class _Overlay(_Feature):
